+++ packages/sdk-shared/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- adding dependabot config generator
 
 ## 0.8.8 - 2021/3/15